requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
python-multipart==0.0.9
pdfplumber==0.11.3
pdf2image==1.17.0
Pillow==10.4.0
pytesseract==0.3.13